fix: scan can be floating point

+++ cbmish.ts
@@ -1431,7 +1431,7 @@ class CbmishConsole {
         value = params.get('scale');
         if (!/^[0-9]+(\.[0-9]+)?$/.test(value))
             return;
-        this.scale = Number.parseInt(value);
+        this.scale = Number.parseFloat(value);
         this.rescale();
     }
